refactor: extract configuration validation to validation service
refs: BLAIS5-4239
